# internal/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

// CreateUserDef is default response for CreateUser operation.
type CreateUserDef struct {
	StatusCode int
}

// GetStatusCode returns the value of StatusCode.
func (s *CreateUserDef) GetStatusCode() int {
	return s.StatusCode
}

// SetStatusCode sets the value of StatusCode.
func (s *CreateUserDef) SetStatusCode(val int) {
	s.StatusCode = val
}

func (*CreateUserDef) createUserRes() {}

// DeleteUserDef is default response for DeleteUser operation.
type DeleteUserDef struct {
	StatusCode int
}

// GetStatusCode returns the value of StatusCode.
func (s *DeleteUserDef) GetStatusCode() int {
	return s.StatusCode
}

// SetStatusCode sets the value of StatusCode.
func (s *DeleteUserDef) SetStatusCode(val int) {
	s.StatusCode = val
}

func (*DeleteUserDef) deleteUserRes() {}

// DeleteUserNoContent is response for DeleteUser operation.
type DeleteUserNoContent struct{}

func (*DeleteUserNoContent) deleteUserRes() {}

// GetUserDef is default response for GetUser operation.
type GetUserDef struct {
	StatusCode int
}

// GetStatusCode returns the value of StatusCode.
func (s *GetUserDef) GetStatusCode() int {
	return s.StatusCode
}

// SetStatusCode sets the value of StatusCode.
func (s *GetUserDef) SetStatusCode(val int) {
	s.StatusCode = val
}

func (*GetUserDef) getUserRes() {}

// GetUserNotFound is response for GetUser operation.
type GetUserNotFound struct{}

func (*GetUserNotFound) getUserRes() {}

// ListUsersDef is default response for ListUsers operation.
type ListUsersDef struct {
	StatusCode int
}

// GetStatusCode returns the value of StatusCode.
func (s *ListUsersDef) GetStatusCode() int {
	return s.StatusCode
}

// SetStatusCode sets the value of StatusCode.
func (s *ListUsersDef) SetStatusCode(val int) {
	s.StatusCode = val
}

func (*ListUsersDef) listUsersRes() {}

type ListUsersOKApplicationJSON []User

func (*ListUsersOKApplicationJSON) listUsersRes() {}

// NewOptNilString returns new OptNilString with value set to v.
func NewOptNilString(v string) OptNilString {
	return OptNilString{
		Value: v,
		Set:   true,
	}
}

// OptNilString is optional nullable string.
type OptNilString struct {
	Value string
	Set   bool
	Null  bool
}

// IsSet returns true if OptNilString was set.
func (o OptNilString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptNilString) Reset() {
	var v string
	o.Value = v
	o.Set = false
	o.Null = false
}

// SetTo sets value to v.
func (o *OptNilString) SetTo(v string) {
	o.Set = true
	o.Null = false
	o.Value = v
}

// IsNull returns true if value is Null.
func (o OptNilString) IsNull() bool { return o.Null }

// SetToNull sets value to null.
func (o *OptNilString) SetToNull() {
	o.Set = true
	o.Null = true
	var v string
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptNilString) Get() (v string, ok bool) {
	if o.Null {
		return v, false
	}
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptNilString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// UpdateUserDef is default response for UpdateUser operation.
type UpdateUserDef struct {
	StatusCode int
}

// GetStatusCode returns the value of StatusCode.
func (s *UpdateUserDef) GetStatusCode() int {
	return s.StatusCode
}

// SetStatusCode sets the value of StatusCode.
func (s *UpdateUserDef) SetStatusCode(val int) {
	s.StatusCode = val
}

func (*UpdateUserDef) updateUserRes() {}

// UpdateUserNotFound is response for UpdateUser operation.
type UpdateUserNotFound struct{}

func (*UpdateUserNotFound) updateUserRes() {}

// Ref: #/components/schemas/User
type User struct {
	ID          int64        `json:"id"`
	Name        string       `json:"name"`
	Description OptNilString `json:"description"`
}

// GetID returns the value of ID.
func (s *User) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *User) GetName() string {
	return s.Name
}

// GetDescription returns the value of Description.
func (s *User) GetDescription() OptNilString {
	return s.Description
}

// SetID sets the value of ID.
func (s *User) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *User) SetName(val string) {
	s.Name = val
}

// SetDescription sets the value of Description.
func (s *User) SetDescription(val OptNilString) {
	s.Description = val
}

func (*User) createUserRes() {}
func (*User) getUserRes()    {}
func (*User) updateUserRes() {}

// Ref: #/components/schemas/UserCreate
type UserCreate struct {
	Name        string       `json:"name"`
	Description OptNilString `json:"description"`
}

// GetName returns the value of Name.
func (s *UserCreate) GetName() string {
	return s.Name
}

// GetDescription returns the value of Description.
func (s *UserCreate) GetDescription() OptNilString {
	return s.Description
}

// SetName sets the value of Name.
func (s *UserCreate) SetName(val string) {
	s.Name = val
}

// SetDescription sets the value of Description.
func (s *UserCreate) SetDescription(val OptNilString) {
	s.Description = val
}

// Ref: #/components/schemas/UserUpdate
type UserUpdate struct {
	Name        OptString    `json:"name"`
	Description OptNilString `json:"description"`
}

// GetName returns the value of Name.
func (s *UserUpdate) GetName() OptString {
	return s.Name
}

// GetDescription returns the value of Description.
func (s *UserUpdate) GetDescription() OptNilString {
	return s.Description
}

// SetName sets the value of Name.
func (s *UserUpdate) SetName(val OptString) {
	s.Name = val
}

// SetDescription sets the value of Description.
func (s *UserUpdate) SetDescription(val OptNilString) {
	s.Description = val
}